+++ pyroSAR/datacube_util.py
@@ -1,6 +1,38 @@
 """
 This (still experimental) module is intended to easily prepare SAR scenes processed
 by pyroSAR for ingestion into an Open Data Cube.
+
+.. code-block:: python
+
+    from pyroSAR.datacube_util import Product, Dataset
+    from pyroSAR.ancillary import find_datasets
+    
+    # find pyroSAR files by metadata attributes
+    archive_s1 = '/.../sentinel1/GRD/processed'
+    scenes_s1 = find_datasets(archive_s1, sensor=('S1A', 'S1B'), acquisition_mode='IW')
+    
+    # define the polarization units describing the data sets
+    units = {'VV': 'backscatter VV', 'VH': 'backscatter VH'}
+    
+    # create a new product
+    with Product(name='S1_GRD_index',
+                 product_type='gamma0',
+                 description='Gamma Naught RTC backscatter') as prod:
+        
+        for dataset in scenes_s1:
+            with Dataset(dataset, units=units) as ds:
+                
+                # add the dataset to the product
+                prod.add(ds)
+                
+                # parse datacube indexing YMLs from product and data set metadata
+                prod.export_indexing_yml(ds, 'yml_index_outdir')
+        
+        # write the product YML
+        prod.write('yml_product')
+        
+        # print the product metadata which is written to the product YML
+        print(prod)
 """
 
 import os